requirements.txt
Django==3.0.2
requests==2.22.0
Pyro4==4.77
python-dotenv==0.10.3